fix test

+++ tasks/all/kholin_k_multidimensional_integrals_rectangle/include/ops_all.hpp
@@ -41,7 +41,7 @@ class TestTaskALL : public ppc::core::Task {
   double I_2n_;
 
   double Integrate(const Function& f, const std::vector<double>& l_limits, const std::vector<double>& u_limits,
-                   const std::vector<double>& h, std::vector<double>& f_values, size_t curr_index_dim, size_t dim,
+                   const std::vector<double>& h, std::vector<double>& f_values, int curr_index_dim, size_t dim,
                    double n);
   double IntegrateWithRectangleMethod(const Function& f, std::vector<double>& f_values,
                                       const std::vector<double>& l_limits, const std::vector<double>& u_limits,
